[fix] Fix problem 347 build error with GCC
1. Integer and unsgined integer comparison warning.

+++ src/medium/347_top_k_frequent_elements.hpp
@@ -13,7 +13,7 @@ namespace top_k_frequent_elements {
 class Solution {
  public:
   std::vector<int> topKFrequent(std::vector<int> &nums, int k) {
-    if (k == nums.size()) {
+    if (static_cast<size_t>(k) == nums.size()) {
       return nums;
     }
